backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-chess==1.999
pytz==2023.3
croniter==2.0.1
psutil==5.9.6
pywin32==306; sys_platform == 'win32'

